requirements.txt
pgeocode
python-dotenv
streamlit
